Refactor tenant API routes for core config list eg/ fetching URL change needed. Fix missing data fetch when listing tenants.
